Added 3D model of jelly with whip and cherry
Cute and scrumptious red jelly with cherries and whipped cream on top!

Just Blender and 15 minutes. While it may have been a quick project, I am very happy with the results and think it turned out adorable.

As a 3D artist, it's important to not only showcase complex and time-consuming projects but also to highlight the beauty in simplicity.

Sometimes, even the smallest of projects can make a big impact!